Fix incorrect left padding issue on tag following float
The return value of `$currblk['padding_left']` can be an integer or a float. Since `is_int` only checks for integers the values that were floats were getting set to 0. Changing it to `is_numeric` covers both integers and floats and brings it inline with the type checking done on lines 739 and 740 of the same file.