Resolve PCRE JIT Stack Limit Error
This PR optimises the Keep With Tables `preg_replace` in the `Mpdf::AdjustHTML()` method so that:

1. It only runs when the `use_kwt` configuration option is enabled
2. The regex match statement has been modified to prevent lookbacks in PCRE

The Problem

In PHP runtimes with PCRE JIT enabled, larger blocks of HTML could trigger a PCRE_ERROR_JIT_STACKLIMIT error if the pattern `<h1></h1><table></table>` isn't in the HTML. This would result in `preg_replace()` returning `null` and the generated PDF would be blank (see https://stackoverflow.com/a/39688067).

As an example of the inefficiencies of the original regex statement can be seen by running the pattern `<(h[1-6])([^>]*)(>(?:(?!h[1-6]).)*?<\/\1>\s*<table)` with the `tests/data/html/long-string-without-kwt-match.html` on https://regex101.com and open up the regex debugger tool they provide, you'll see a total of 205 603 steps for zero matches.

The Solution

First, let me say this is a stop-gap solution and we should eventually look at using a real HTML DOM Parser.

In the interim, the solution I came up with was to only run this particular regex if the configuration calls for it. The `use_kwt` configuration is disabled by default in mPDF so this will increase rendering performance for those using the defaults.

Then I've re-written the expression to be more efficient and not do unnecessary backtracking (see https://stackoverflow.com/a/2146912). This was checked using regex101.com against the original HTML string, resulting in 47 159 steps.

Just in case an error does occur due to `preg_replace()` we will ignore the results and continue the process. I had considered throwing an exception here instead of continuing, but I'd match rather generate a PDF that may not keep the headings with the tables rather than a PDF that is blank.